Extract prompt template detection into microcrate
